Move useApplyCode to hooks directory
